Move checkpoint manager into own file
